separate all of the logic from DOM stuff
