+++ src/app/main.js
@@ -1,14 +1,32 @@
+import { game, gameboard } from './dom'
+
+let player1 = '';
+let player2 = '';
+
 export const gamePlayers = (() => {
+    let current = true;
     const playerFactory = (name, symbol, score) => ({ name, symbol, score });
 
-    const players = (name, symbol) => {
-        return playerFactory(name, symbol, 0);
+    const players = (name1, name2) => {
+        player1 = playerFactory(name1, 'X', 0);
+        player2 = playerFactory(name2, 'O', 0);
+        game.play(player1, player2);
+        game.score(player1, player2);
     }
 
-    return { players }
+    const switchPlayers = (player1, player2) => {
+        if (current === true) {
+          current = false;
+          return player1;
+        }
+        current = true;
+        return player2;
+      };
+
+    return { players, switchPlayers }
 })();
 
-export const gameBoardLogic = (() => {
+export const gameboardLogic = (() => {
 
 const victoryArr = [[1, 2, 3], [4, 5, 6], [7, 8, 9],
     [1, 4, 7], [2, 5, 8], [3, 6, 9],
@@ -16,34 +34,47 @@ const victoryArr = [[1, 2, 3], [4, 5, 6], [7, 8, 9],
     
 let gameboardArr1 = [];
 let gameboardArr2 = [];
+let counter = 0;
 
     const checkWin = (player, gameboardArr) => {
     
         for(let i = 0; i < victoryArr.length; i += 1) {
+            console.log(gameboardArr)
+            console.log(victoryArr[i])
             if (gameboardArr.includes(victoryArr[i][0])
                 && gameboardArr.includes(victoryArr[i][1])
                 && gameboardArr.includes(victoryArr[i][2])) {
                 player.score += 1;
-                game.score();
+                game.score(player1, player2);
                 return `${player.name} wins!`;
-            } else if (counter === 9 && result.textContent === '') {
-                game.score();
-                return "It's a draw!";
             }
         };
+        if (counter === 9) {
+            return "It's a draw!";
+        }
     }
     
     const populateArr = (position, symbol) => {
         if (symbol === 'X') {
-        gameboardArr1.push(Number(position) + 1);
-        writeResult(player1, gameboardArr1);
+            gameboardArr1.push(Number(position) + 1);
+            gameboard.writeResult(player1, gameboardArr1);
         } else {
-        gameboardArr2.push(Number(position) + 1);
-        writeResult(player2, gameboardArr2);
+            gameboardArr2.push(Number(position) + 1);
+            gameboard.writeResult(player2, gameboardArr2);
         }
     };
 
-    return { checkWin, populateArr}
+    const resetArr = () => {
+        gameboardArr1 = [];
+        gameboardArr2 = [];
+        counter = 0;
+    }
+
+    const increaseCounter = () => {
+        counter += 1;
+    }
+
+    return { checkWin, populateArr, resetArr, increaseCounter }
 })();
